requirements.txt
Django==4.2.7
pdfkit==1.0.0
sqlparse==0.4.4
typing-extensions==4.8.0
djangorestframework